Remove subclasses of mod:SemanticArtefact
The subclasses `mod:Taxonomy`, `mod:Thesaurus`, and `mod:Terminology`
don't add any value to MOD but only confusion because they don't come
with additional semantics and they are not strictly defined. In fact,
the use of names such as "taxonomy" and "terminology" for specific types
of semantic artifcats differs widely among disciplines. Other
vocabularies of types of knowledge organization systems exist to be used
if needed. For instance BARTOC.org uses the NKOS KOS Types vocabulary,
based on research in Knowledge Organization.